Project: rev 216 Note- almost finished still some buffer methods out of wack and I think some got named wrong

+++ api-rs/src/main/java/net/runelite/rs/api/RSBuffer.java
@@ -34,7 +34,7 @@ byte readByteNeg();
 
 @Import("readUnsignedByteSub")
 @Override
-byte readUnsignedByteSub$api();
+int readUnsignedByteSub$api();
 
 @Import("readUnsignedByte")
 @Override
